Minor update - setup script
Minor update to the setup script to be able to define the IP and host the flask app runs under.

Also added a force refresh button, so when you add a group you do not need to wait until the next refresh windows, you can push it then and there.